Switched to urlsafe IDs for Devices, instead of numeric IDs. * NDB re-uses child IDs if the parent is different. (But not on localhost for some reason.) This caused a bug because I fetched the wrong Device when I was trying to edit another person's Device. * This required a change to a few functions, plus the returned Device JSON, plus the route for Devices (alphanumeric now instead of numeric). * Other small bug fixes after running the tests on the Appspot cloud. * Removed User.get_by_id(). I didn't realize it earlier, but NDB models already have this function. Though I need to pass in a Long explicitly to the built-in version.

+++ UserHandlers.py
@@ -74,7 +74,7 @@ class UserHandler(webapp2.RequestHandler):
 
     def put(self, user_id):
 
-        user = User.get_by_id(user_id)
+        user = User.get_by_id(long(user_id))
 
         if not user:
             send_error(self.response, 404)
@@ -104,7 +104,7 @@ class UserHandler(webapp2.RequestHandler):
 
     def delete(self, user_id):
 
-        user = User.get_by_id(user_id)
+        user = User.get_by_id(long(user_id))
 
         if not user:
             send_error(self.response, 404)
